Fix multishipping address issue

+++ Plugin/Magento/Quote/Api/CartRepositoryInterface.php
@@ -149,11 +149,12 @@ class CartRepositoryInterface
         }
         else {
             $result[1] = $buyerModel->getBuyerId();
-            $verifier = $this->buyerApi->getBuyer($buyerModel->getExternalIdentifier());
-            if ($buyerModel->getBuyerId() != $verifier->getId()) {
-                // remove faulty record
-                $this->buyerRepository->delete($buyerModel);
-                $result[0] = true;
+            if ($verifier = $this->buyerApi->getBuyer($buyerModel->getExternalIdentifier())) {
+                if ($buyerModel->getBuyerId() != $verifier->getId()) {
+                    // remove faulty record
+                    $this->buyerRepository->delete($buyerModel);
+                    $result[0] = true;
+                }
             }
         }
 
@@ -172,8 +173,9 @@ class CartRepositoryInterface
         $buyer_id = $this->buyerApi->createBuyer($external_identifier, $display_name);
 
         if ($buyer_id == Gr4vyBuyer::ERROR_DUPLICATE) {
-            $buyer = $this->buyerApi->getBuyer($external_identifier);
-            $buyer_id = $buyer->getId();
+            if ($buyer = $this->buyerApi->getBuyer($external_identifier)) {
+                $buyer_id = $buyer->getId();
+            }
         }
 
         return $buyer_id;
